Use bsr instead of division for performance.