@rhysd 2021-09-08
Thank you. I made [a small benchmark suite](https://github.com/rhysd/bit_combi_iter/blob/main/benches/bench.rs) and confirmed the performance improvement as follows:

Before:

```
test u32_3 ... bench:      35,852 ns/iter (+/- 2,694)
test u64_3 ... bench:     384,282 ns/iter (+/- 19,453)
test u8_1  ... bench:          32 ns/iter (+/- 1)
test u8_3  ... bench:         357 ns/iter (+/- 18)
test u8_5  ... bench:         349 ns/iter (+/- 3)
```

After:

```
test u32_3 ... bench:      10,341 ns/iter (+/- 606)
test u64_3 ... bench:      83,871 ns/iter (+/- 9,084)
test u8_1  ... bench:          13 ns/iter (+/- 0)
test u8_3  ... bench:         132 ns/iter (+/- 8)
test u8_5  ... bench:         137 ns/iter (+/- 5)
```

It was around 2.46x ~ 4.58x speed up.